refactor: move RunnerLifecycleService, RunnerPollerConformances, RunnerEditDraft ext to Core (#1618)
- Add RunnerBarCore/Runner/RunnerLifecycleService.swift (public struct + public init)
- Add RunnerBarCore/Runner/RunnerPollerConformances.swift
- Add RunnerBarCore/Runner/RunnerEditDraft+ProductionLoad.swift
- Tombstone all three source files in RunnerBar/Runner/

Note: RunnerEditDraft+ProductionLoad depends on #1616 (RunnerProxyStore → Core).
Build will succeed once #1616 is merged.
